// src/shared/types.d.ts
interface Image {
  name?: string;
  link?: string;
  src: string 
  // | StaticImageData;
  alt: string;
  // tags:Array[string]
}
// ...
interface Storelist {
  stores: Array<Image>;
}
// ...
interface review{
  appname:string,
  review:string,
  storename:string
}